+++ packages/adhere-mobile-ui-spin/es/Spin.d.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import type { SpinProps } from './types';
+declare const Spin: React.NamedExoticComponent<SpinProps>;
+export default Spin;