+++ packages/adhere-mobile-ui-spin/es/types.d.ts
@@ -1,0 +1,9 @@
+import type { ToastProps } from 'antd-mobile/es/components/toast/toast';
+import { CSSProperties } from 'react';
+export type SpinProps = Omit<ToastProps, 'getContainer' | 'content' | 'visible' | 'duration'> & {
+    className?: string;
+    style?: CSSProperties;
+    spinning?: boolean;
+    text?: ToastProps['content'];
+    zIndex?: number;
+};