+++ packages/adhere/script/build.js
@@ -101,6 +101,7 @@ const namedMap = new Map([
   ['@baifendian/adhere-mobile-ui-prsl', 'MobilePRSL'],
   ['@baifendian/adhere-mobile-ui-quick-range-date', 'MobileQuickRangeDate'],
   ['@baifendian/adhere-mobile-ui-popovermenu', 'MobilePopoverMenu'],
+  ['@baifendian/adhere-mobile-ui-spin', 'MobileSpin'],
 ]);
 
 /**