+++ packages/adhere/src/mobile-spin/index.js
@@ -1,0 +1,2 @@
+import Model from '@baifendian/adhere-mobile-ui-spin/es';
+export default Model;